refactor: crée un fichier PianoTrackerContext
