+++ site/source/components/PianoAnalytics/PianoTrackerContext.ts
@@ -1,0 +1,8 @@
+import { createContext, useContext } from "react"
+import { PianoTracker } from "./PianoTracker"
+
+export const PianoTrackerContext = createContext<PianoTracker | null>(null)
+
+export function usePianoTracker(): PianoTracker | null {
+	return useContext(PianoTrackerContext)
+}
